profile and nearby page are more interacting now

+++ frontend/src/components/Profile.js
@@ -2,10 +2,14 @@ import React, { useState, useEffect, useContext } from "react";
 import "bootstrap/dist/css/bootstrap.min.css";
 import EditProfile from "./EditProfile";
 import axiosClient from "./axiosClient";
-import { NavLink } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 import "./components.css";
 import { AuthContext } from "./context/AuthProvider";
 import Card from "react-bootstrap/Card";
+import { FaEdit } from "react-icons/fa";
+import { AiFillDelete } from "react-icons/ai";
+import { FaSearchLocation } from "react-icons/fa";
+import { Roller } from "react-awesome-spinners";
 
 function Profile() {
   const [show, setShow] = useState(false);
@@ -14,39 +18,77 @@ function Profile() {
   const [profile, setProfile] = useState({});
   const [location, setLocation] = useState({});
   const { deleteProfile } = useContext(AuthContext);
+  const [checkLocation, setCheckLocation] = useState(true);
+  const [loading, setLoading] = useState(false);
+  const navigate = useNavigate();
 
   useEffect(() => {
-    navigator.geolocation.getCurrentPosition(function (position) {
-      setLocation({
-        type: "point",
-        coordinates: [position.coords.latitude, position.coords.longitude],
-      });
-    });
-
+    setLoading(true);
     axiosClient
       .get("/users/profile")
       .then((res) => {
+        // console.log(res.data);
         setProfile(res.data);
+        setLoading(false)
+        if (res.data.location.coordinates) {
+          setCheckLocation(false);
+        } else {
+          if (!navigator.geolocation) {
+            alert("Geolocation is not supported by this browser.");
+          } else {
+            navigator.geolocation.getCurrentPosition(function (position) {
+              setLocation({
+                type: "point",
+                coordinates: [position.coords.latitude, position.coords.longitude],
+              });
+              axiosClient
+                .put("/users/profile", {
+                  location: {
+                    type: "point",
+                    coordinates: [
+                      position.coords.latitude,
+                      position.coords.longitude,
+                    ],
+                  },
+                })
+                .then((res) => {
+                  console.log(res.data)
+                  setProfile(res.data);
+                  if (res.data.location.coordinates) {
+                    setCheckLocation(false);
+                  } else {
+                    setCheckLocation(true);
+                  }
+                })
+                .catch((err) => console.log(err));
+            });
+          }
+        }
       })
       .catch((err) => console.log(err));
+
+  
   }, []);
 
-  const sendLocation = async () => {
-    try {
-      const result = await axiosClient.put("/users/profile", { location });
-      setProfile(result.data);
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
+  // const sendLocation = async () => {
+  //   try {
+  //     const result = await axiosClient.put("/users/profile", { location });
+  //     setProfile(result.data);
+  //   } catch (error) {
+  //     console.log(error);
+  //   }
+  // };
   const deleteAccount = () => {
-    deleteProfile();
+    const response =  window.confirm('Are you sure you want to delete your account')
+    console.log(response)
+    if (response == true) {
+      deleteProfile();
+    }
   };
 
   return (
     <div className="row ">
-      <div className="profile my-5 d-flex flex-wrap justify-content-center p-5">
+      {loading ? <div className="d-flex flex-colum justify-content-center my-5"><Roller color="#0e4d4d" /></div> : <div className="profile my-5 d-flex flex-wrap justify-content-center p-5">
         <Card className="col-6 col-md-3 text-center">
           <Card.Img
             variant="top"
@@ -57,7 +99,7 @@ function Profile() {
           <Card.Body className="d-flex flex-column justify-content-around">
             <Card.Title>{profile.name}</Card.Title>
             <button className="btn btn-success text-white" onClick={handleShow}>
-              Edit Profile
+              <FaEdit /> Edit Profile
             </button>
           </Card.Body>
         </Card>
@@ -72,16 +114,18 @@ function Profile() {
           </div>
           <div className=" d-flex justify-content-between">
             <button className="btn btn-danger" onClick={deleteAccount}>
-              Delete Account
+              <AiFillDelete /> Delete Account
             </button>
-            <NavLink to="/people-nearby" className="nav-link">
-              <button className="btn btn-success" onClick={sendLocation}>
-                See people nearby
-              </button>
-            </NavLink>
+            <button
+              onClick={() => navigate("/people-nearby")}
+              title="Your current location will be used to see people nearby"
+              className="btn btn-success"
+              disabled={checkLocation}
+            >
+              See people nearby <FaSearchLocation />
+            </button>
           </div>
         </div>
-
         <EditProfile
           setShow={setShow}
           show={show}
@@ -89,49 +133,11 @@ function Profile() {
           profile={profile}
           setProfile={setProfile}
         />
-      </div>
+      </div>}
+      
     </div>
   );
 }
 
 export default Profile;
 
-{
-  /* <div className="row profile my-5">
-      <div className="col-12 m-5 d-flex flex-wrap ">
-        <div className="col-6 flex-wrap">
-          <img src={profile.pic} alt="profilepic" id="profilepic" />
-        </div>
-        <div className="col-6 flex-wrap">
-          <p>Name
-          <h3>{profile.name}</h3>
-          </p>
-          <p>Description
-          <h5>{profile.description}</h5>
-          </p>
-          <p>Hobbies
-          <h5>{profile.hobbies}</h5>
-          </p>
-          <p>German Level
-          <h5>{profile.germanLevel}</h5>
-          </p>
-        </div>
-      </div>
-      <div className="d-flex  justify-content-between m-5">
-        <button className="btn btn-success text-white" onClick={handleShow}>
-          Edit Profile
-        </button>
-          <button className="btn btn-danger" onClick={deleteAccount} >Delete Account</button>
-        <NavLink to="/people-nearby" className="nav-link">
-          <button className="btn btn-success" onClick={sendLocation} >See people nearby</button>
-        </NavLink>
-      <EditProfile
-        setShow={setShow}
-        show={show}
-        handleClose={handleClose}
-        profile={profile}
-        setProfile={setProfile}
-      />
-      </div>
-    </div> */
-}
